Refactor out the preprocessing from ASR into common
