+++ examples/tts/test_tts_infer.py
@@ -26,7 +26,7 @@ import torch
 
 from nemo.collections.asr.metrics.wer import word_error_rate
 from nemo.collections.asr.models import EncDecCTCModel
-from nemo.collections.asr.parts import parsers
+from nemo.collections.common.parts.preprocessing import parsers
 from nemo.collections.tts.models.base import SpectrogramGenerator, Vocoder
 from nemo.utils import logging
