+++ nemo/collections/common/parts/preprocessing/collections.py
@@ -19,7 +19,7 @@ from typing import Any, Dict, List, Optional, Union
 
 import pandas as pd
 
-from nemo.collections.asr.parts import manifest, parsers
+from nemo.collections.common.parts.preprocessing import parsers, manifest
 from nemo.utils import logging
